Example10MarkerFile {Map2NCBI}	R Documentation

Example Marker File with 10 Markers on BTA 1

Description

This file contains marker map placement based on Btau 6.1 and is meant to be used as an example of the MapMarkers function.

Usage

data(Example10MarkerFile)

Format

The format is: chr "Example10MarkerFile"

Details

Markers are SNP from BovSNP50 assay by Illumina, Inc.(San Diego, CA). These markers were used in a Nellore-Angus crossbred population described in the reference below.

References

Riley, D.G., Welsh Jr., T.H., Gill, C.A., Hulsman, L.L., Herring, A.D., Riggs, P.K., Sawyer, J.E., Sanders, J.O., 2013. Whole genome association of SNP with newborn calf cannon bone length. Liv. Sci. 155: 186-196.

Examples

data(Example10MarkerFile)

[Package Map2NCBI version 1.1 Index]