Map2NCBI-package {Map2NCBI}	R Documentation

Mapping Markers to the Nearest Genomic Feature

Description

Allows the user to generate a list of features (gene, pseudo, RNA, CDS, and/or UTR) directly from NCBI database for any species with a current build available. Option to save downloaded and formatted files is available, and the user can prioritize the feature list based on type and assembly builds present in the current build used. The user can then use the list of features generated or provide a list to map a set of markers (designed for SNP markers with a single base pair position available) to the closest feature based on the map build. This function does require map positions of the markers to be provided and the positions should be based on the build being queried through NCBI.

Details

Package: 	 Map2NCBI

Type: 	 Package

Version: 	 1.1

Date: 	 2015-11-12

License: 	 GPL (>= 2)

This package can be used as a two part process with the "GetGeneList" function followed by the "MapMarkers" function. See individual function documentation for more information.

Author(s)

Lauren L. Hulsman Hanna and David G. Riley

Maintainer: Lauren Hanna <Lauren.Hanna@ndsu.edu>

References

Hulsman Hanna, L. L., and D. G. Riley. 2014. Mapping genomic markers to closest feature using the R package Map2NCBI. Livest. Sci. 162:59-65.

See Also

Functions: GetGeneList & MapMarkers

Examples

#See individual function documentation for applied examples.

[Package Map2NCBI version 1.1 Index]